refactor(web): move scope types and storage utils to lib/leaderboard-scope.ts
Extract ScopeSelection, Organization, Team types and
loadScopeFromStorage / saveScopeToStorage utilities from
scope-dropdown.tsx into a dedicated lib/leaderboard-scope.ts module.
scope-dropdown.tsx re-exports them for backward compatibility.
use-leaderboard-scope.ts now imports directly from the new module.
